requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
diffusers==0.25.0
transformers==4.36.0
torch==2.1.0
torchvision==0.16.0
accelerate==0.25.0
xformers==0.0.22
peft==0.7.1
supabase==2.0.2
python-multipart==0.0.6
pillow==10.1.0
numpy==1.24.3
requests==2.31.0
python-dotenv==1.0.0
pydantic==2.5.0
opencv-python==4.8.1.78
compel==2.0.2